requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS
python-dotenv==1.0.0

# Geospatial
shapely==2.0.2

# HTTP Client (para llamar a otros microservicios)
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

